dnsdist: Revert the now unneeded changes made to dnsdist-healthchecks
